refactor: extract IComputeRaysOutFactory
